separate game logic module
