+++ README.md
@@ -1,0 +1,27 @@
+# Kralı Bul (Find the King)
+
+This repository contains a small Python game with a Tkinter GUI. The goal is to find the hidden king tile on a 5x5 board using numbered cards. The rules can be customized or expanded later.
+
+## Requirements
+
+- Python 3.x with Tkinter support (usually comes pre‑installed on Windows and many Linux distributions).
+
+## Running the Game
+
+1. (Optional) Create and activate a virtual environment:
+   ```bash
+   python3 -m venv venv
+   source venv/bin/activate
+   ```
+2. Run the script:
+   ```bash
+   python Find-the-king.py
+   ```
+
+A window should open showing the board and card selection buttons.
+
+## Notes
+
+- The rules of the game can be updated in this README once finalized.
+- The GUI lives in `Find-the-king.py` while the core logic is in
+  `game_logic.py`.
